requirements.txt
beautifulsoup4==4.12.3
black==24.4.2
Django==5.0.4
djangorestframework==3.15.2
psycopg2-binary==2.9.9
pytest==8.3.2
requests==2.27.1
python-decouple==3.8
pytest==8.3.2
google-api-core==2.19.1
google-auth==2.33.0
google-cloud-core==2.4.1
google-cloud-translate==3.16.0
googleapis-common-protos==1.63.2
grpc-google-iam-v1==0.13.1
grpcio==1.65.4
grpcio-status==1.65.4
cachetools==5.4.0
